Upgraded AllowedConversions from i32 to i128 to increase award precision without changing Borsh versions yet.

+++ masp_proofs/src/circuit/convert.rs
@@ -144,15 +144,15 @@ fn test_convert_circuit_with_bls12_381() {
 
     let tree_depth = 32;
 
-    for _ in 0..10u32 {
+    for _ in 0..10u128 {
         let i = rng.next_u32();
         let spend_asset = AssetType::new(format!("asset {}", i).as_bytes()).unwrap();
         let output_asset = AssetType::new(format!("asset {}", i + 1).as_bytes()).unwrap();
         let mint_asset = AssetType::new(b"reward").unwrap();
 
-        let spend_value = -(i as i32 + 1);
-        let output_value = i as i32 + 1;
-        let mint_value = i as i32 + 1;
+        let spend_value = -(i as i128 + 1);
+        let output_value = i as i128 + 1;
+        let mint_value = i as i128 + 1;
 
         let allowed_conversion: AllowedConversion = (ValueSum::from_pair(spend_asset, spend_value)
             .unwrap()
